+++ src/utils/customTypes.ts
@@ -1,0 +1,5 @@
+export type ApiResponse = {
+	success: boolean;
+	message: string;
+	data: any;
+};
