+++ web/src/utils/constants.ts
@@ -19,6 +19,9 @@ export const reactionsUrl = `${apiBaseUrl}/api/reactions`;
 export const artUrl = `${apiBaseUrl}/api/art`;
 export const achievementsUrl = `${apiBaseUrl}/api/achievements`;
 export const totalsUrl = `${apiBaseUrl}/api/profile/totals`;
+export const friendRequestsUrl = `${apiBaseUrl}/api/profile/friendrequests`;
+export const addFriendUrl = `${apiBaseUrl}/api/profile/addfriend`;
+export const respondToFriendRequestUrl = `${apiBaseUrl}/api/profile/friendrequest`;
 export const defaultAvatarUrl =
 	'https://64.media.tumblr.com/660a46fe7a5825d898e36fc8c240685b/bd38759d70dd2985-ab/s400x600/05eb60dacdbb47e946713a8aaa048274dc176cb7.png';
 export const twitterUsername = '@matthughes2112';
